requirements.txt
asgiref==3.7.2
colorama==0.4.6
coverage==7.4.2
Django==5.0.2
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==23.2.1
inflection==0.5.1
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.1
pytest-cov==4.1.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
